Add remaining response tests
I had to revert to mocking the higher level `_wait_for_response` call
instead of the `asyncio.Future` call b/c when running all the tests in
parallel different mocking calls started overwriting each other.
I would still prefer to mock the `recv_multipart` call to the socket in
order to be as close to the metal as possible. Alas, that is for an
other day.