Group spents with receives
This is trickier than it sounds. The spents have as the value field a checksum
of a hash. This is home baked.
In this patch you'll also see an experimentation with the @staticmethod
annotation. I'm not sure what I find more horrible, the module level public
methods or the @staticmethod annotation (which allows the method to be called
on the Class *and* an instance of the Class).